SAtt-UNet_Repo/requirements.txt
torch
torchvision
numpy
pandas
matplotlib
scikit-learn
tqdm
Pillow
seaborn
torchinfo
thop
ptflops
